[WIP] Replace GetBufferDescriptor with memref::extract_strided_metadata
The following patch is a draft of the direction we are going with
memref::extract_strided_metadata in "MLIR core".  Essentially, this
operator has the same semantic as GetBufferDescriptor except that the
return `base_buffer` is in the memref space, not the iree one. To close
this gap, we introduce a `get_buffer_pointer` operation that is just a
wrapper around the returned `base_buffer` memref.

The current patch is not funtional for a lot of cases because the core
MLIR logic is still missing the simplification patterns for alloc op
(and some other simplifications that were not performed with
GetBufferDescriptor.)
(It does work in a handful of cases though!)

The idea here is just to demonstrate what it will look like in the
future. (I put comments where the code should be moved (though my
suggestion may not be the correct ones!) and what should be removed.)