Fix mmt4d_test: actually dont use unsupported instructions.
This was regressed by #10485.

Also, report the skipped/ran status on stdout not stderr, because GTest
output goes there so that's how we get it consistently interleaved.

Also, fix a numerical bug in the i8mm kernel introduced in #10476.
It was caught by mmt4d_test, not sure how I didn't notice.