Make iree-benchmark-suites buildable with some backends disabled. (#10508)
Note: the reason why iree-test-deps wasn't failing similarly was that
`iree_single_backend_generated_trace_runner_test` and
`iree_check_single_backend_test_suite` were already performing a similar
check, but they do it on the `DRIVER`, not on the `TARGET_BACKEND`. They
can do so because these targets are built in the device build dir, where
`IREE_HAL_DRIVER_*` variables are set and drivers are built, because
drivers are part of the runtime, so they're built in the device build
dir.

Unlike those targets, `iree-benchmark-suites` is built in the host build
dir, so we shouldn't enable or disable benchmarks based on
`IREE_HAL_DRIVER_*` here. Instead, what is relevant (and to the point of
fixing the module compile error that caused me to look into this) is a
check on `IREE_TARGET_BACKEND_*`.

Possibly these could be unified by switching also `iree-test-deps` to
this `IREE_TARGET_BACKEND_*` check.