Keras layers that do not support fully dynamic shapes are tested with a dynamic batch size only.
Keras currently calls `set_shape` based on the shape of the `Input` that is defined. However, this logic in Keras is being removed as it is incorrect in some scenarios.

Because of this, the IREE "dynamic_dims" unit tests were actually running with a fully static shape and not exercising operations with a dynamic shape. Testing with dynamic shapes uncovered the fact that some layers do not support it.