Adding `--expected_output=` flag to iree-run-module. (#10433)
This allows for super basic blackbox testing of compiled modules by way
of iree-run-module that does not require any input changes like
iree-check-module does. The flag accepts the same values as
`--function_input` (including npy files) in addition to `(ignored)` for
when a value is not of interest.

When the flag is provided the normal output printing is disabled and
instead a comparison runs and prints detailed information about any
diffs, for example:
```
[FAILED] result[0]: element at index 49 (0) does not match the expected (123); expected that the view is equal to contents of a view of 5x1x10xf32
  expected:
5x1x10xf32=[[0.743973 ... 0 0 123]]
  actual:
5x1x10xf32=[[0.743973 ... 0 0 0]]
```

The matching utilities are intended to be wrapped in gtest C++ matchers
so that they can be used in iree-check-module and other tools but that
is left for future work. This separation is required so that
iree-run-module and other tools using the comparison utilities don't
need to depend on gtest. It also means that users authoring their own
test tools for bare metal/etc can easily use the comparison features.

There are three flags like `--expected_f32_threshold=` (+f16/f64) that
allow for basic threshold overrides. Full customization of the equality
comparisons that do smarter things (ULP, integer ranges, etc) are also
left for future work and what's used here matches iree-check-module's
current behavior.

Fixes #10376.