Add simple Python webserver for GitHub runner health check
For now, it literally always returns an HTTP 200 response. All this does
is confirm that the startup script has completed and nothing has killed
the health server. Unlike the current approach that traps an ERR signal
and shuts down the instance, this is extensible to more sophisticated
examination of whether the GitHub runner is actually functioning. This
puts the recreation of the instance in the hand of the instance group,
which makes it more visible and doesn't work at cross-purposes with the
infrastructure we're using.