Add ops to set and unset tensor encoding to capture layout changes.
The `iree_linalg_ext.pack` and `iree_linalg_ext.unpack` operations
represent the change in data layout by explicitly changing the rank of
the tensors. This might be useful for code-generation, but higher up
the stack results of these operations will then need special tiled
versions of operations like `linalg.matmul` and `linalg.conv_*`.
Instead, the ops added here represent the layout change as a tensor
encoding without changing the logical shape of the tensor. These set
and unset encoding operations then can be lowered to pack and unpack
operations that explicitly introduce the higher dimensional tensors
within a dispatch.

The encoding attribute used can be richer, but for now they are just
an enum with semantic meaning that can be used as required. Over time
these encoding attributes can be modified to be more explicit in
describing the layout change to perform.