shark model testing for iree (#9748)
Attempts to address #9537 

Has the same workflow inputs as validate_and_publish_release.yml, but I
wasn't sure what the intended place in the existing workflows this was
meant to plug into. All it currently does is download the generated
packages from the build_package.yml workflow and use them to run shark
model zoo. Happy to configure to spec.

I think the custom runners are still up in the air per #9432 , so this
will have to be edited later to account for how those are used.

Co-authored-by: mariecwhite <mariewhite@google.com>