Modify `pack` op iteration space description to make it amenable to Tiling. (#10569)
The innermost loop dimensions of the `pack` operation deal with the packing the data into a tile. These loops cannot be tiled. The loops that iterate over the different data tiles of the output can be tiled. Change the iteration space description to only specify the loops that iterate over the data tiles of the output. The intra-data tile loops are generated as part of the scalar implementation of the `pack` operation.