Publish release pip index as part of website (#10594)
This is incorporated into the docs pipeline and served on the website.
This isn't really the ideal place to serve it and we should probably
create a new subdomain and host a PEP 503 compliant index there. But
this will unbreak people for now.

Note that this includes a change to make docs publication run on forks.
I did this initially to enable testing, but I actually think it's a
reasonable change to make. The workflow is fast and cheap and this
makes testing it easier. GitHub already provides a way for people to
disable all actions or specific actions in their repository and I think
it makes more sense to let them handle it that way. I am open to being
convinced otherwise though. The fork most likely to be doing something
else with their gh-pages branch is Nod, and I've checked that they
aren't (nothing here: https://nod-ai.github.io/SHARK/)

Another attempt at https://github.com/iree-org/iree/pull/10581 and
https://github.com/iree-org/iree/pull/10585

Supersedes https://github.com/iree-org/iree/pull/10583.

Part of https://github.com/iree-org/iree/issues/10479.

Tested:
https://github.com/GMNGeoffrey/iree/actions/runs/3150453503 ->
https://gmngeoffrey.github.io/iree/pip-release-links.html

I didn't test the release trigger because doing an actual release is
very time-consuming.

skip-ci