Add a code size and no-C++ check on Android.
This generates as CI error if any C++ makes it into the IREE runtime, at
least the part of it that is linked into a minimal C program calling
some `iree_vm_invoke` (adapted from samples/custom_module).

This also adds some binary size monitoring: the current sections size of
that binary is ~ 397,000 bytes, so, allowing ourselves 20% growth, an
arbitrary limit is set at 480,000 bytes. When we hit that limit, we will
get a CI error inviting us to decide whether that's normal and we just
want to bump the limit.

This matters specifically on Android because Android does not
treat the C++ runtime as a "system library" --- it's up to each application to
distribute its own, and the CMake default, which we are using, is to
statically link it. See https://developer.android.com/ndk/guides/cpp-support.

To make this worse, even though IREE C++ code is built with `-fno-rtti` and
`-fno-exceptions`, the libc++ that we are statically linking against in the
Android NDK is the same as everyone else's --- just because we have those
compilation flags doesn't change what we link against. As a result, a single
`operator new` is enough to drag in a large amount of exception-handling,
stack-unwinding, demangling code.

For example, on Android/AArch64, `iree-run-module` contains 120 kB of
C++ symbols.