[LLVMGPU] Workaround wrong signed extension of buffer offsets
Buffers can currently only have 32bits unsigned offset but we
mistakently signe extend them to 64bits for 64bits targets. In order to
prevent getting a negative offset we force trunc the offset to 32bits
and drop the higher bits.

The proper fix would be to support zero extended casting for i32 to
index. The proper fix will be tracked here:
https://github.com/iree-org/iree/issues/10615