Skip zero-filling accumulators, let ukernels overwrite (#10645)
When the accumulator argument to a `linalg.matmul` or `linalg.mmt4d` is
filled with constant zeros, we can drop that fill and instead tell the
`vmvx.matmul` or `vmvx.mmt4d` to overwrite the output buffer instead of
loading and accumulating into the existing values.

This wasn't the top performance bottleneck --- slow packing code was ---
but it was a good thing to work on independently while
@MaheshRavishankar and @hanhanW are bringing up the new packing ops. And
it still was significant: here this makes a 128x128, `i8` matmul over
20% faster. Okay, to put this into some perspective, we are talking
about going from 1.8 Gop/s to 2.2 Gop/s (peak is 760 Gop/s, the mmt4d
ukernel part does 730 Gop/s so all the remaining slowness is the
packing/unpacking code. To put that in perspective, current llvmcpu does
100 Gop/s), so it's still very slow with all that packing code going
down the scalar VM bytecode path, but it's good to get this out of the
way so that when we turn attention to packing code, the impact of that
will not be diluted in slow zero-filling code.