NFC: Add static method to `Pack` op to get result shape.
Add a static class method to `Pack` operation that allows getting the
shape of the expected result. Such methods helps to lower into pack
operations that is destination passing style.
Also add a custom `build` method to `Pack` ops to hide a lot of bioler
plate.
Fix a lit test along the way.