Relax verifier restrictions for `outer-dims-perm`. (#10752)
Originally `inner-dims-pos` and `outer-dims-perm` had to agree on the
dimension to tile and interchange, meaning that the vector must have the
same elements. This patch relaxes the restriction and allows
`outer-dims-perms` to target any tiled loop.

We also fix a bug in the interchange routine. `outer-dims-pos`
represents the new position for the tile loops. For example, given the
dimension [1, 55, 56, 2] and the interchange vector [0, 3, 1, 2] we
expect: [1, 2, 55, 56]. Dimension `2` in the old position `3` is now
expected at position `1.