[iree-dialects][transform] drop unnecessary handle associations
Canonicalization in the canonicalized sequence transform op may replace
payload operations associated with the transform dialect handles.
Post-canonicalize reassociation is fragile and may fail. To make this
less likely, drop any association that are no longer necessary, i.e., if
the operand is no longer used in the sequence or elsewhere or if the
result is never read. This is a more robust but IREE-specific
alternative to the obsolete functionality in upstream MLIR that wouldn't
associate unread results.

Closes #10785