[spirv] Add initial I64 emulation pass (#10682)
The pass is conditionally enabled when the SPIR-V target does not
declare I64 capabilities. The `demote-i64-to-i32' needs to be disabled
for Wide Integer Emulation (WIE) to have any i64 ops to emulate. These
default will be changed once the WIE implementation is robust enough.

Locally, I tested this by changing the `demote-i64-to-i32` default to
false and all tests passed.

Issue: #8702