Enable e2e `linalg_ext.pack` test on VMVX + ukernel (#10798)
This is in preparation for actually writing a microkernel.

This is only pack for now, not unpack, because of
https://github.com/iree-org/iree/issues/10796 .

I had to introduce a new compiler command-line flag,
`--iree-vmvx-enable-microkernels-decompose-linalg-generic`, defaulting
to `true` (the current behavior), so that I could set it to `false` in
this test. The problem is that when `--iree-vmvx-enable-microkernels` is
passed, we use a DecomposeLinalgGenericPass. The `pack.mlir` test here
contains `linalg.generic` which, when decomposed by this pass, are split
into a chain of multiple `linalg.generic` with temporary buffers in
between. These temporary buffers are `memref.alloca` in the IR,
resulting in compilation failures:

```
error: failed to legalize operation 'memref.alloca' that was explicitly marked illegal
```