Add support for generalizing and vectorizing linalg_ext.pack op. (#10734)
To vectorize the pack op, we generalize the pack op to tensor.pad + linalg.generic ops and kick in upstream generic vectorizer.

In the PR, we require outer dimensions of results to be all 1s. The restriction can be relaxed, but that requires us to add special vectorization pattern for expand_shape. The other way to support the cases is to tile outer dims, make them be all 1s and kick in generalization and generic vectorizer.