Adding CPU import conversion and ordinal assignment machinery. (#10865)
During conversion to the LLVM dialect any `func.call` to an extern
function or a function emitted during conversion will be lowered into an
import call with a placeholder ordinal. This uses a technique like
executable constants whereby imports get an attributed global
placeholder for the ordinal that is assigned in the new
LLVMCPUAssignImportOrdinals pass after linking. During serialization the
import table built during assignment is used to declare the imports on
the executable library data structure.

The automatic conversion comes courtesy of @nicolasvasilache and should
handle most basic arg/result packing.
Minor fixes were required in the runtime for MSVC compatibility with
embedded ELFs.

Fixes #7504.