Relax isolatedFromAbove requirement in the ListenerGreedyPatternrewriter
All known uses of applyPatternsAndFoldGreedily are on op->getRegions().
Modify the API to always operate on Operation* and use that as the root op to determine when to avoid adding new ops to the workList.

Note: This may need to be reverted in the future but I conjecture this restriction is nt really load-bearing and may bea remnant of
multi-threaded parallel pass pipeline to which we do not subscribe with the transform dialect.