[spirv] Only unify aliased resources on Apple platforms
The UnifyAliasedResourcePass is actually only necessary for targeting
Apple GPUs via MoltenVK, where we need to translate SPIR-V into MSL.
The translation has limitations--no support of aliased resources.

There is a hidden bug in the UnifyAliasedResourcePass yet to be
identified. So for now limiting it to only Apple platforms to avoid
further disruption.