Tightening ukernel common definitions (#10950)
* Complete rewrite of the big comment at the start of ukernel/common.h.
Please read carefully as I might be saying things that are incorrect.
* Ukernel local types prefixed to disambiguate from standard types, e.g.
`int32_t` --> `iree_ukernel_int32_t`.
* This caught a place in `vmvx/module.c` that now requires an explicit
cast from `unsigned long*` to `unsigned long long*`. I know these are
always the same 64bit stuff in practice, but it's worth an explicit
cast.
* Actually no longer including various platform headers in ukernels. In
ukernel/common.h, we were including iree/base/attributes.h, which was
including target_platform.h, which was including windows.h etc.
* Now ukernels that need a RESTRICT keyword can use
IREE_UKERNEL_RESTRICT, forked from IREE_RESTRICT.
* We can almost, but not quite, simply use `restrict`, since ukernels
are C code. The problem is that C++ code may need to include the
declarations. At least, the macro can make it easy to experiment with
what's the impact of restrict. And it leaves the door open to C++
kernels if someone really wants that in the future.