Fix parsing trailers when they're the only thing in the PR body (#10974)
We already generally treat the PR description as a commit description
with title and body in this way (e.g. in our [squash and] merge commit
messages). I could've added any dummy title string, but this seemed
more understandable. We could also write our own parsing instead of
relying on interpret-trailers, but I think it's good to use a standard
tool.

Tested:
https://github.com/iree-org/iree/actions/runs/3364878153/jobs/5579737086