Add support for generalizing unpack op when outer dim sizes are all 1s.  (#10990)
It generalizes the unpack op into extract_slice + copy + insert_slice
ops. Because the outer dim sizes are all 1s, we're able to extract the
inner tiles from input, apply the transposition, and insert it back to a
tensor. It enables the vectorization for unpack op.

For outer dim sizes that are not all 1s, we have to update the upstream
`scf::tileUsingSCFForOp` method to accept multi operations created
during tiling. That would be the next step after landing this commit.

The commit also moves `computeInterchangeFromDimPos` to utils.