Some support for variadic tile size generation
Add a unit attribute to TileToForeachThreadAndWorkgroupCountRegionOp
indicating that the tile size handle points to a list of tile sizes
rather than individual tile size. This allows for a variadic list of
tile sizes to be passed. Hook this up with the preexisting tile size
generation. Additionally, wrap the tile size generation and the tiling
itself into a regular sequence so canonicalization doesn't run and
doesn't remove tile sizes before they are being used. Note that running
canonicalization after a "match" or another transformation that doesn't
mutate IR is a waste of cycles by itself.