Adding a makefile to the dynamic_shapes sample. (#11093)
This is just intended as an example of how to construct a makefile that
can compile and link a binary against a build of the IREE runtime
unified library. Users in other build systems or with their own cmake
utilities can use it to see what linking looks like. There are a lot of
caveats with this approach and not using IREE as a cmake subproject is
still YMMV.