Add a pass to materialize the encoding during executable translation. (#11056)
Under the data tiling approach, at the Flow level an iree_linalg_ext.encoding attribute is added to tensors to capture the need for a data layout transformation. During executable translation this encoding + information about backend can be used to decide how the encoding is to be materialized (using pack and unpack operations). This PR adds a pass to materialize the encoding for the CPU backends. It uses patterns from a similar pass in IREE::LinalgExt::LinalgExtDialect, in addition to adding legalizations for hal.interface.binding,
flow.dispatch.tensor.load/store and type conversion for !flow.dispatch.tensor type. To facilitate this

- A new class called MaterializeEncodingTypeConverter, that derives from TypeConverter, is added that is meant to be used with patterns that implement the materialization.
- A new class called MaterializeConversionTarget, that derives from ConversionTarget, is added that is meant to be used with the Dialect Conversion framework to drive the materialization.
- Both the above classes rely on a callback being implemented by the users of these class that given the encoding return the MaterializeEncodingInfo object that has everything needed to materialize the pack/unpack operations.
- This patch also adds the lowering for the flow.dispatch.workgroup_count_from_set_encoding_op needed to compute the number of workgroups correctly, based on the workload that was captured at the Flow level.

With these changes a linalg.matmul can be compiled with IREE using the --iree-flow-enable-data-tiling flag.