[vulkan] Use unknown triple as the default (#10529)
This commit adds version/extension/capability values for a
"unknown-unknown-unknown" target triple, and use it as the default if no
explicit ones are specified. This triple is very conservative; it only
picks version/extension/capability that is very popular and using Vulkan
required limits. So that the generated binary is expected to run without
problem on most Vulkan devices. With it we can detech from SwiftShader
specific characteristics, like, for subgroup size, we can use a more
resonable default than SwiftShader's 4.