Add LinalgExt canonicalization patterns to tile+distribute pass. (#11143)
The cast operations can be created during tiling. We have to fold it
into the tiled op if possible. The pattern for LinAlg dialect is
included, but the pattern for LinalgExt dialect is not included. This
commit adds the pattern to the pass.

Fixes https://github.com/iree-org/iree/issues/11038