Introduce a transform.do_not_dce_operands
The new op is introduced in LinalgExt for now as it cannot be dynamically registered into the transform dialect atm.
With this op, canonicalized_sequence does not DCE the SSA values introduced by ConfigExtractPartOp.

This is necessary to allow the next part of the config to be extracted without either DCE or dropped attributes under transforms limiting us.

With the existing ListenerGreedyPatternRewriteDriver.cpp, the new op would be be DCE'ed (!!!).
So we also adapt the code to its recent upstream evolution.