Add support for extracting second-level tiling from config.
This commit allows using the extracted config values for the second level tiling.
In the process, we add the ability to dynamically extract list of operations from pdl::OperationType values, both here and upstream.

This greatly simplifies manipulating pdl::Operation that capture lists of operations, interspersed with regular Values and Attr.