Translate flow.tensor.empty to stream.tensor.empty. (#11131)
I believe this matches the expected behavior of tensor.empty as
@silvasean pointed out in #8717. I've added the
`--iree-flow-zero-fill-empty-tensors=` flag to allow an override if we
notice user programs that assume the contents are zero-filled.

On ESRGAN this saves nearly 1GB of memset(0)'s per invocation:
  Original fill count: 627 size: 2004094976 (2GB)
  New fill count: 351 size: 1117008896 (1.1GB)

Workaround for #6972. We still want generalized tracking but this will
cover concats (the most common producers of this pattern).
Progress on #7729.