Convert HAL::DescriptorTypeAttr to proper attributes (#11202)
This commit converts HAL::DescriptorTypeAttr to be a proper attribute,
with its own attribute storage; so not an InegerAttr derived one
anymore. This is in preparation to be used as MemRef "memory space".

Along the way fixed an issue in `walkAttributeStorage` where we aren't
guard against unknown custom attributes--it will just be silently
ignored, as the functor won't be called at all so `anyFailed` won't be
updated. Additionally check the status of `walkAttributeStorage`.

Progress towards #11183