Address use-def violations arising due to index computation when rewriting tensor.load/tensor.store ops.
When creating `flow.dispatch.tensor.load`/`flow.dispatch.tensor.store`
based on already existing operations (during
`-iree-codegen-tile-and-distribute-to-workgroups` and
`-iree-convert-to-destination-passing-style`), using the `Value`s for
the offsets, sizes and strides directly can lead to SSA use-def
violations. So clone the computation of these values. These are
expected to be a few operations that eventually get CSEed.  Add helper
methods to do this in `Codegen/Transforms.h` and
`Codegen/Transforms.cpp` for general use.