Add basic structured op matcher with fluent API
This plugs into the m_Foo/matchPattern approach but allows for
finer-grain specification of the structured op properties specifically.
The intention is to further extract this properties into the transform
dialect and/or PDL so they are specified using a fixed, though
extensible, set of predicates rather than C++ lambdas.