Matcher API for inputs and inits/outputs
Add a simple recurisve matching API for inputs and inits of a structured
op that may be fed by another structured op. This currently expect a
concrete position and additional work is needed to support "any
position" or wildcarding. This API follows other fluent functions
instead of special-casing input matchers to be constructor operands,
i.e. one is expected to write `consumer.input(0, producer())` rather
than `consumer(producer())`. This is expected to scale better in
presence of wildcards and input/init mixes where depending on the
position of the argument in the argument list is undesirable.