ASan tests: account for multi-driver Lit tests (#11242)
This PR reenables all tests that were disabled under ASan!

The LSAN failures that motivated the current exclude list should all be
fixed by #11206.

Yet, simply dropping the exclude-list doesn't work, because some of the
entries in that list are Lit tests that exercise multiple drivers in a
way that's encoded as `//RUN` comments within the test source, opaque to
the build and test system --- there was no way to tell which of those
tests involved Vulkan.

It would be nice to only disable LSAN for the Vulkan part of these
tests, but this isn't feasible as things are currently structured, with
the multiple driver runs conflated withing a single Lit test source.

At first I introduced a new tag/label, `multi-driver-including-vulkan`,
to label those tests, so that `build_and_test_asan.sh` could disable
LSAN accordingly.

Then I thought, how does this scale? Will there be a `multi-driver-XYZ`
tag for all combinations of (possibly multiple?!) drivers?

Then I thought, let's just use the existing `driver=vulkan` tag for
this, giving it "multiple choices possible" semantics. Actually, "tags"
are meant to work that way; perhaps the only thing about that tag that
makes it somewhat surprising that it would work that way is the `=`
sign: it might seem strange to have `driver=X` and `driver=Y` tags
together on the same rule. Then again maybe we'll get used to it, and
maybe precisely having such tags side by side (as enforced by
`buildifier`) will make it clear that it does have such semantics.