Shut down runner VM if startup fails (#11250)
I was planning to do this as health check, which integrates with the
actual MIG infrastructure, but that has ended up being a laborious
process due to Google security rules and I still haven't gotten it to
work. In the meantime, we've been having frequent issues with server
errors from the GitHub actions control plane, which mean that the
runners are never registered and the instances just sit around
counting as "running" according to the autoscaler.

This is a really simple method to just shut down the runners if setup
fails. We should also add retries and stuff, but this works for now and
should hopefully avoid us ending up with half our instance pool just
sitting idle.

Tested:
I added a non-deterministic failure in the setup script (see reverted
commit) and confirmed that some of the instances initially shut down,
but eventually all came online as runners.

skip-ci: Runner startup configuration not tested by CI.