Create explicit compiler C API libraries and enable dynamic linking on all platforms.
This is the first step of a multi-part sequence that will:

* Rebuild the compiler/API directory into its final form.
* Move the compiler python API to its correct location (bindings/python).
* Enable dynamic linking project wide of all compiler tools.

This first step leaves the following undone:

* Python API stays where it is (there are a lot of fiddly path dependencies there and I opted to handle that in isolation).
* Sets everything up in an API2 which will eventually be renamed to API.
* Just gets the stub of what is needed for a delay loaded embedding API -- the full thing will follow and be integrated into PJRT for Jax and TensorFlow serving on the OSS side.
* The project local lld is still building static, even though the API includes it in the shared library. Was just trying to keep the patch size down vs there being a fundamental problem.

In the new state, there is a libIREECompiler.so.0 (or IREECompiler.dll on Windows) which contains both the in-process APIs and the tool entry-points. The following tools have been reworked to link against the shared library:

* iree-compile
* iree-lld (Python side only)
* iree-opt
* iree-mlir-lsp-server

In addition, some API tests which were doing expensive static linking now link dynamically. All said, this saves ~4-5x multi-gigabyte compiler tool linking steps and should be a pretty substantial build iteration improvement. I wasn't originally going to include iree-opt and iree-mlir-lsp-server in the shared build, but based on an experiment, they contribute negligible size (really just a few extra source files when linked together), and the productivity and usability savings (i.e. we can now distribute those at negligible cost) made me decide it was worth it.

I don't know what to do about iree-run-mlir. I think I can refactor it into something based on the compiler API that then links the runtime. But that is for later.

I took some steps to ensure this will work on Windows and MacOS but haven't finished verifying.