Support reduction+elementwise in transform dialect strategies
Add a handwritten and a generated transform dialect schedule to support
codegen for reduction+elementwise targeting CUDA. Unlike standalone
reduction, this needs additional fusion into the trailing elementwise.
The transform dialect generated on-the-fly differs when the elementwise
is present. This should be eventually expressed in the dialect itself,
but we are currently lacking the necessary ops.