Adding IREE_LINK_COMPILER_SHARED_LIBRARY cmake flag. (#11335)
This is enabled by default and tools link in the fancy shared library.
When disabled the tools link in the static library version. This can be
useful when wanting to build standalone binaries. It's also useful to
work around MSVC's link.exe issues with ilk files >4GB that we run into
with the shared library enabled.