Support distribution to more than 3 dimension at the workgroup level
So far we were not distributing higher level dimensions for operations
with more than 3 parallel dimensions. This delinearize the z dimension
for case with more than 3 dimension and allow better distribution if the
kernel config asks for it.
Also change the default config of llvmgpu and spirv backend to take
advantage of that.