Use C++ callbacks to match reductions inside the transform dialect (#11390)
Instead of relying on the fragile matching of all payload operations of
the same kind and splitting them in the fixed number of singleton
buckets, use the C++ callback to match the reduction structure. Now, the
exact same callback is used to find the case where the transform
dialect-based pipeline applies and to identify the ops that are
transformed later on. This is still suboptimal as the matching is done
twice, but further integration is left to future work. So are
integration tests via C++ builder API for the transform dialect.