[spirv] Fix C matrix promotion with bufferization allocations (#11418)
For cases like linalg.matmul + arith.extf, bufferization inserts an
allocation because of the different element type of those two ops. Here
we already have the C matrix in workgroup memory; there is no need to
promote it again. Though we need to make sure to use the proper marker
so that we can insert barriers and enable following steps to distribute
to SIMT threads.

Co-authored-by: Thomas <thomasraoux@google.com>