Rename FG_WRITE_ACCESS_TOKEN back to WRITE_ACCESS_TOKEN
The transition to the new token style was successful and I"ll be
removing the old-style token, so there's no need for the disambiguating
prefix. I've already changed the value of the secret to be new
fine-grained token.

Fixes https://github.com/iree-org/iree/issues/11479

skip-ci: Does not affect presubmit workflows