Emit better reduction schedule from JIT (#11498)
Emit the reduction schedule splitting the reduction directly to
"scf.for" rather than to a generic over inputs with expanded shape. This
allows us to ensure the generation of vector loads and stores, and to
target the vector size appropriate to the input. Also plug in the
matcher into the schedule instead of the fragile reliance on the order
of operations.