[WIP] A proxy server for allow GCE instances to delete themselves
GCE Managed instance groups don't have any good way to handle
autoscaling for long-running workloads. With the autoscaler configured
to scale in, instances get only 90 seconds warning to shut down. So we
set the autoscaler to only scale out and have the VMs tear themselves
down when they're down with their work. But anything that brings down
the VM other than a delete call to the instance group manager API makes
the VM get considered "unhealthy", which means it gets recreated in
exactly the same configuration, regardless of any update or autoscaling
settings. Making the correct API call requires broad permissions on the
instance group manager, which we don't want to give the VMs. To scope
permissions to individual instances, this proxy service makes use of
instance identity tokens to allow an instance to make a call only to
delete itself.

See
https://cloud.google.com/compute/docs/instances/verifying-instance-identity

Note that http.server is not in general fit for production, but our very
limited usage of BaseHTTPRequestHandler here, not serving any files and
not doing our own parsing of request input, does not present any
security concerns.