Speed up filter constant folding
The existing approach took on the order of minutes
for filters with ~1k input and ~1k output channels.
This patch reduces that to the order of seconds.

The primary change is we allocate space for a float
array that contains the constant filter and convert
from APFloat to float. Then, in a separate loop we
perform the folding operating on floats and only
store the result as an APFloat (with appropriate
precision).