Add constraints on the number of payload ops that we must match to av… (#11568)
…oid false positives while matching.

These extra constraints are necessary because of the layering IREE
imposes on us: the dispatch regions are already pre-formed and we must
match **exactly** (best effort..) to avoid leaving dangling payload IR
in the dispatch that is not transformed (and leads to catastrophic
performance bugs or even miscompiles).

A future more robust system will let us form our own dispatches and we
won't need to be as strict on matching **exactly**.

This PR makes the assumption that TilingInterface ops are exactly the
payload ops that we must not miss.

This is best effort for now and if anything else must not be missed, it
should also be added.

With this PR, we avoid polluting cases that we should not pollute.

This is one separable step towards integration that requires additional
steps to be useful. Still this can be sliced out and C++ tests should be
added independently.

Deferring to @ftynse for refactoring and testing this slice.