Generate accessors for public global variables (#10531)
As part of importing ml_program, generate accessors for public global
variables. This just handles the simple case get and (if mutable) set
methods that take or return the full value. The naming convention is
configurable to avoid implicit linkage until/if convention is fixed.

If accessors are not specified, then the generation is skipped.