Fix performance bugs and add a benchmarking stub
This revision introduces a 3 stage stategy for mapping reductions
along with a simple heuristic that finds the right tile sizes and vector sizes as a function
of the reduction size, the max number of threads and the warp shuffle size.

This provides a first set of tunable knobs for this strategy that can evolve into something more general in the future.

A simple benchmarking stub file is added that can be used with `sed`-based commands.
It is currently unclear where the control for such a benchmarking should live and what relevant sizes would be.