Cloud Run Function for instance self-deletion (#11610)
A Cloud Functions proxy enabling GCE VMs in a Managed Instance Group to
delete themselves. GCE Managed instance groups don't have any good way
to handle autoscaling for long-running workloads. With the autoscaler
configured to scale in, instances get only 90 seconds warning to shut
down. So we set the autoscaler to only scale out and have the VMs tear
themselves down when they're down with their work. This is the approach
suggested by the managed instance group team:
https://drive.google.com/file/d/1XlwxF_0T7pUnbzhL5ePDoW-Q3GAaLO11. But
anything that brings down the VM other than a delete call to the
instance group manager API makes the VM get considered "unhealthy",
which means it gets recreated in exactly the same configuration,
regardless of any update or autoscaling settings. Making the correct
API call requires broad permissions on the instance group manager,
which we don't want to give the VMs. To scope permissions to individual
instances, this proxy service makes use of instance identity tokens to
allow an instance to make a call only to delete itself (see

https://cloud.google.com/compute/docs/instances/verifying-instance-identity).

This makes use of the GCP Cloud Functions serverless offering. It's
another level of abstraction on top of Cloud Run, where you don't even
need to create your own docker container.

I wrote unit tests for this in addition to testing the server locally,
but I don't think it's worth adding them to the CI right now because
they come with an entire server framework of dependencies and this
isn't a part of the project I expect to see active development.
Probably more worthwhile would be finding another home for this as a
standalone project separate from IREE.

skip-ci: not tested by CI