Extra transform interpreter targeting with attributes
Introduce an additional targeting mechanism to the transform dialect
interpreter pass. This mechanism makes the pass walk the operation it is
running on to find another operation, marked with the
`transform.iree_tag` string attribute with the specified value, that is
used as payload root. Similarly, the pass walks either the operation it
runs on or the transform IR module when it was loaded externally to find
an opreation marked with `transform.iree_tag` string attribute with the
specified value that is used as the container for transform ops.

This connects better to IREE's deeply nested attribute-controlled pass
pipeline machinery. It has the immediate benefit of enabling repro
generation, also added in this patch. The repro can be dumped to
standard debug output stream (when running interactively) or to a file
using the `-debug-only=iree-transform-dialect-dump-repro` or
`-debug-only=iree-transform-dialect-save-repro` command line flag,
respectively. The repro command is printed to the standard debug stream.