[Transform][Debug] Extra transform interpreter targeting with attributes (#11629)
Introduce an additional targeting mechanism to the transform dialect
interpreter pass.

The purpose of this addition is to greatly improve debugging velocity by
navigating more smoothly between:
  1. C++ builders that produce transform IR
2. Transform IR textual manipulation and local application with the
interpreter.
3. Re-injection of manually altered transform IR in the system to run
end-to-end and prove/disprove a hypothesis.

Note that steps 2 and 3 above create an interpreted sandbox in which one
can very efficiently modify strategies and produce IR at any point
without either recompiling the compiler or sifting through walls of text
from `print-ir-after-all`.

This mechanism makes the pass walk the operation it is running on to
find another operation, marked with the `transform.iree_tag` string
attribute with the specified value, that is used as payload root.
Similarly, the pass walks either the operation it runs on or the
transform IR module when it was loaded externally to find an operation
marked with `transform.iree_tag` string attribute with the specified
value that is used as the container for transform ops. Note that this
does not control individual transformations. It is also not expected to
be used in normal lowering pipeline.

This connects better to IREE's deeply nested attribute-controlled pass
pipeline machinery. It has the immediate benefit of enabling repro
generation, also added in this patch. The repro can be dumped to
standard debug output stream (when running interactively) or to a file
using the `-debug-only=iree-transform-dialect-dump-repro` or
`-debug-only=iree-transform-dialect-save-repro` command line flag,
respectively. The repro command is printed to the standard debug stream.

Co-authored-by: Nicolas Vasilache <nicolas.vasilache@gmail.com>