Revert "Support transform dialect debug instructions for iree-compile (#11654)"
This reverts commit 565e8f51a10c87595b25f01ddf734f700b54d77c.

This seems incomplete and I can't follow what it is doing. It seems to be exercising a workflow that is pretty non-standard and, at a minimum, needs explanation. Most likely, this should be carried in a branch and/or have more review so there can be a better shared understanding. The `findFuncWithDispatchOfSymbol` function has a lot of TODOs and cuts across incorrect layers: we can't be doing that in the production code.

Given holidays and outages, just rolling back and can pick up when folks are back.