Make Bazel Cuda work for dependencies on IREE (#11666)
Somehow the way Label is evaluated differs if your Bazel macro is in
the main repository, in which case it gets evaluated relative to the
BUILD file's repository vs if the macro is being used as part of an
external repository, in which case it is evaluated relative to the
macro's repository (which is the documented behavior).

Hardcoding the repo alias would break anyone using this function from
an external repository who doesn't call the IREE repo "iree_core".
Removing `clean_dep` would break anyone using `iree_c_embed_data` from
an external repository. The latter isn't really supported, but it
doesn't make things any harder in the end.