Use local SSDs instead of ramdisks on runners
While there is plenty of RAM on the CPU runners, the GPU runners are
pretty limited. We unfortunately need to mount a silly number of disks
for CPU instances. I looked at making use of all of them as a single
logical volume as described in
http://cloud/compute/docs/disks/add-local-ssd#formatmultiple,
but formatting took quite a while and 375GB is plenty of space.