Test transform matching via iree-opt
Add a mechanism for testing the result of transform dialect-related
matchers as follows. A new transform op emits remarks at the location of
payload ops associated with the given handle. This op is used to each
handle produced by `match_callback` to mark the matched payload ops.
Thus iree-opt-based tests can check the matcher directly with the
`-verify-diagnostics` flag.