Rank-reducing / hoisting / vectorization interplay repro.
This PR shows that swapping ApplyPatternsOp and VectorizeOp is currently necessary for hoisting to occur properly.
This is related to collapse/expand around scf.for that do not fold away nicely and in turn block hoisting on tensors.

```
(benchmark-transform-create -r  ../iree-samples/transform_dialect/benchmark_linalg_reductions.stub.mlir   reduction_2d_static 1234567 42)
```

takes a 4x hit when hoisting does not occur.

@matthias-springer