Add debug output to transform dialect matchers (#11674)
This makes it easier to understand failures-to-match. It required
splitting the implementation of some function templates into a public
template and private non-template part with the latter emitting the
debug output to avoid polluting the header with debug-related includes.
Note that this starts approaching the point where turning matchers into
operations in a dialect would be preferable.