Fix out of bounds access in `emitGPUGroupReduction`.
Previously, we would emit IR such as:
```
%2 = gpu.thread_id  x
...
%15 = arith.remui %2, %c32 : index
...
%17 = memref.load %alloc_11[%15] : memref<3xf32, 3>
```
with `gpu.thread_id x` ranging in `[0 .. 95]`.

This performs a clear out-of-bounds read for threads `[3..31] + 32 * {0, 1, 2}`.
Interestingly, the error does not seem to result in crashes in simpler divisble IR such as line 42 in:
https://gist.github.com/nicolasvasilache/004f9f23ba0bde9b3517e5f0d6fb6989

It does result in crashes in more complex, non-divisible cases, such as line 79 in:
https://gist.github.com/nicolasvasilache/e145eab16df75d494eceac34e45331ce

It is unclear atm whether this is a case that the nvcc compiler managed optimized away or whether something more
advanced is going on.

This PR additionally confines the lane index to be in the bounds of the memref:

```
%2 = gpu.thread_id  x
...
%15 = arith.remui %2, %c32 : index
%16 = arith.minui %15, %c2: index
%17 = memref.load %alloc_11[%16] : memref<3xf32, 3>
```

With this change, crashes are now gone.