Modify aggressive fusion heuristics to not create ops where reduction ops also return intermediate results.
Currently under aggressive fusion, elementwise parallel ops are fused
with consumer that are reductions. When the parallel op has other uses
the fused op also returns the reduced and non-reduced values. This
results in operations that are reductions returning results that have
shape similar to the entire iteration space of the op. Such ops are
hard to handle in the backend.  This patch avoids creating such ops,
but adapts the fusion heuristics to still pull these ops into the same
dispatch.