Default instance group updates to 'opportunistic replace' (#11727)
This used 'proactive refresh' previously to get around the bug/feature
in MIGs that recreation of VMs because they are unhealthy doesn't apply
any pending updates. Now that we have our workaround to have the VMs get
deleted through the MIG API, updates actually will get applied.
Replacing the VM is necessary for several update operations (e.g. stuff
with disks and policies). Our most common changes are just metadata
changes so they can be applied with refresh, but they actually don't
take
effect until the VM restarts, so it's misleading to show the update as
having been applied.

Also fixed some comments and help text in this and the instance_deleter.