Introduce Collapse Pass (#11713)
This is a follow-up PR. #11295 had already implemented the collapsing
logic for `linalg.generic` operations in `dispatch.region`. This PR
extracts that logic in a new Collapse Pass.

The new pass performs the following steps. The first and third steps
here are new compared to the previous one as we are creating the Pass.
1. Finds `linalg.generic` that has no producer in `dispatch.region`,
which means there will only be one `linalg.generic` in the region.
2. Collapse dimensions of the tensors in `linalg.generic` if possible
3. Generates new workload region for the collapsed Op and new
`dispatch.region`

PR also removes collapse flag and changes a test that was using
accordingly.