Purge split reduction transformation from APIs
Split reduction is subsumed by tiling reductions to foreach_thread which
is more general and can reduce temporary storage to arbitrarily small sizes.

For now we keep the SplitReduction matcher as it is used for more general testing
and the current test cannot be easily rewritten until the BubbleCollapseExpand maintains handles.