Remove the need for split_reduction match callback
This match callback was introduced to replace the original "match all
generics/fills + split" approach, which itself required re-matching
because of:

  1. `apply_patterns` invalidating the parent function handle and thus
     all handles to nested ops;
  2. `fuse_into_containing_op` not preserving the order of payload
     operations associated with the handle (and rather returning the
     results in the order in which they ended up being fused).

Remedy 1. by marking `apply_patterns` as only reading and not consuming
the function handle, thus avoiding its invalidation. This is made
possible by the implementation of `apply_patterns` using the rewrite
driver with tracking that preserves the handles even when the associated
operations were affected.

Remedy 2. by fusing one op at a time to obtain the handle to the op
after fusion. This hit an issue with canonicalization effectively
fusing a `fill` op into the loop when it is directly consumed, confusing
the fusion op. Work around it by performing fusion inside a
non-canonicalized sequence.

This allows us to remove the `split_reduction` matcher and related code,
as well as to simplify the transformation script by removing mutliple
`take_first` that became redundant.